+++ lib/notify.ts
@@ -1,8 +1,0 @@
-import swal from 'sweetalert';
-
-const config = {
-  className: 'next-smrt-theme',
-};
-
-export const notify = (msg: string) =>
-  swal('Notification!', msg, 'success', config).catch((e) => e);