+++ lib/utils.ts
@@ -1,4 +1,10 @@
-export const reduceToIDs = (acc: number[], val: { id: number }) => {
-  acc.push(val.id);
-  return acc;
-};
+export function createRandomKey(idLength: number = 5) {
+  let text = '';
+  const possible =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+
+  for (let i = 0; i < idLength; i++)
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+
+  return text;
+}