+++ next.config.js
@@ -1,7 +1,10 @@
 module.exports = {
+	compiler: {
+		styledComponents: true,
+	},
 	env: {
 		API_URL: process.env.API_URL || "http://next-smrt.csprance.localhost/api",
 		API_TOKEN: process.env.API_TOKEN || "API_TOKEN"
 	},
-	webpack5: true,
+	reactStrictMode: true,
 };