+++ pages/about.tsx
@@ -1,14 +1,35 @@
-import Box from '@material-ui/core/Box';
+import { Box } from '@mui/material';
 import { GetStaticProps, NextPage } from 'next';
 import Head from 'next/head';
+import Image from 'next/image';
 import * as React from 'react';
 
 import Layout from '../components/Layout';
 
-interface Props {
+export const getStaticProps: GetStaticProps = async (context) => {
+  // This joke will be baked in to the static page and then rebaked every 30 seconds if visited
+  try {
+    const response = await fetch('https://api.chucknorris.io/jokes/random');
+    const joke = await response.json();
+    return {
+      props: {
+        joke: joke.value,
+      },
+      revalidate: 30,
+    };
+  } catch (e) {
+    return {
+      props: {
+        joke: 'Error Fetching a cool Chuck Joke Sorry.',
+      },
+      revalidate: 30,
+    };
+  }
+};
+
+export const AboutPage: NextPage<{
   joke: string;
-}
-export const AboutPage: NextPage<Props> = ({ joke }) => {
+}> = ({ joke }) => {
   return (
     <Layout>
       <Head>
@@ -23,10 +44,12 @@ export const AboutPage: NextPage<Props> = ({ joke }) => {
         justifyContent={'center'}
         style={{ padding: 10, textAlign: 'center' }}
       >
-        <img
-          style={{ maxWidth: '100%' }}
+        <Image
           src="/static/site-image.png"
           alt="next-smrt-logo"
+          width={840}
+          height={222}
+          layout={'intrinsic'}
         />
         <h1>Next-SMRT</h1>
         <h2>
@@ -52,27 +75,6 @@ export const AboutPage: NextPage<Props> = ({ joke }) => {
       </Box>
     </Layout>
   );
-};
-
-export const getStaticProps: GetStaticProps = async (context) => {
-  // This joke will be baked in to the static page and then rebaked every 30 seconds if visited
-  try {
-    const response = await fetch('https://api.chucknorris.io/jokes/random');
-    const joke = await response.json();
-    return {
-      props: {
-        joke: joke.value,
-      },
-      revalidate: 30
-    };
-  } catch (e) {
-    return {
-      props: {
-        joke: 'Error Fetching a cool Chuck Joke Sorry.',
-      },
-      revalidate: 30
-    };
-  }
 };
 
 export default AboutPage;