+++ store/todo/types.ts
@@ -1,9 +1,6 @@
 // All the types associated with the state from the db
-
 export type Todo = {
   id: number;
   completed: boolean;
-  todoText: string;
+  text: string;
 };
-
-export type State = Todo[];